+++ src/utils/getTagsforPostCollection.js
